requirements.txt
asgiref==3.5.2
cycler==0.11.0
Django==4.1
django-queryset-csv==1.1.0
djangorestframework==3.13.1
fonttools==4.37.1
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
padasip==1.2.2
pandas==1.4.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
si==0.1.1.post1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
unicodecsv==0.14.1
